Built initial APK; Api not connected or no data retrieved; New initial data in home. Built initial APK for app, release key created with credentials alias:codechallenge and both passwords = codechallenge. Changed android app icon. Disable landscape mode via Android Studio.
Changes in async function, return an empty array if no data is retrieved 
and if a network is not detected. Error is catched in views by comparing 
if response is empty.

Added two new arrays of data to display in flat lists in home. Added 
most popular and top rated animes.